requirements_all.txt
appdirs==1.4.4
audioread==2.1.9
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
colorama @ file:///tmp/build/80754af9/colorama_1607707115595/work
cycler==0.10.0
dataclasses @ file:///tmp/build/80754af9/dataclasses_1614363715916/work
decorator @ file:///tmp/build/80754af9/decorator_1621259047763/work
entrypoints==0.3
idna==3.2
ipykernel @ file:///C:/ci/ipykernel_1596206775157/work/dist/ipykernel-5.3.4-py3-none-any.whl
ipython @ file:///C:/ci/ipython_1593446240034/work
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
jedi==0.17.0
joblib==1.0.1
jupyter-client @ file:///tmp/build/80754af9/jupyter_client_1630690655192/work
jupyter-core @ file:///C:/ci/jupyter_core_1612213536848/work
kiwisolver @ file:///C:/ci/kiwisolver_1612282446297/work
librosa==0.8.1
llvmlite==0.36.0
matplotlib @ file:///C:/ci/matplotlib-suite_1613408055530/work
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
nest-asyncio @ file:///tmp/build/80754af9/nest-asyncio_1613680548246/work
numba==0.53.1
numpy @ file:///C:/ci/numpy_and_numpy_base_1603480701039/work
olefile==0.46
packaging==21.0
pandas==1.1.5
parso @ file:///tmp/build/80754af9/parso_1617223946239/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow @ file:///C:/ci/pillow_1625663293114/work
pooch==1.5.1
prompt-toolkit @ file:///Users/ktietz/demo/mc3/conda-bld/prompt-toolkit_1629146841569/work
pycparser==2.20
pydub==0.25.1
Pygments @ file:///tmp/build/80754af9/pygments_1629234116488/work
pyparsing @ file:///home/linux1/recipes/ci/pyparsing_1610983426697/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytz==2021.1
pywin32==228
PyYAML==5.4.1
pyzmq @ file:///C:/ci/pyzmq_1628276569449/work
requests==2.26.0
resampy==0.2.2
scikit-learn==0.24.2
scipy==1.5.4
six @ file:///tmp/build/80754af9/six_1623709665295/work
sklearn==0.0
SoundFile==0.10.3.post1
threadpoolctl==2.2.0
timm==0.4.12
torch==1.9.0
torchaudio==0.9.0
torchvision==0.10.0
tornado @ file:///C:/ci/tornado_1606942379977/work
traitlets==4.3.3
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1624965014186/work
urllib3==1.26.6
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
wincertstore==0.2
